requirements.txt
gradio
langchain
langchain-community
langchain-experimental
requests
python-dotenv
openai